Added code that embeds source filename into music.
When music is decoded, embedded filename is automatically extracted, and data is saved to that file.  Known issue: if the source filename contains path information, this could cause problems when opened on a different machine that has a different directory structure.  I intend to remedy this by scrubbing all path information from the filename prior to embed.